s2: Move DynamicGraph::execute to DynamicPipeline::push()
